Use float32 for bias and tau parameters

+++ fla/layers/swattn_triton.py
@@ -88,12 +88,12 @@ class SWAttention(nn.Module):
         # Lazy Attention 的可学习参数
         # 位置 bias: [num_heads, max_bias_length]
         # 距离范围 [0, max_bias_length)，与原始实现一致
-        # Use float32 to avoid gradient underflow with bfloat16 training
-        self.learnable_bias_diagonals = nn.Parameter(torch.zeros(self.num_heads, self.max_bias_length, dtype=torch.float32))
+        self.learnable_bias_diagonals = nn.Parameter(torch.zeros(self.num_heads, self.max_bias_length))
         nn.init.normal_(self.learnable_bias_diagonals, mean=0.0, std=1e-3)
 
         # Elastic-Softmax 的 τ 参数: [num_heads]
-        # Use float32 to avoid gradient underflow with bfloat16 training
+        # Use float32 explicitly to avoid gradient underflow with bfloat16 training
+        # Initialize to -1.0 as per paper, but consider -0.1 if dead heads persist
         self.tau = nn.Parameter(torch.full((self.num_heads,), -1.0, dtype=torch.float32))
 
         self.rotary = RotaryEmbedding(dim=self.head_dim, base=self.rope_theta)
@@ -161,11 +161,12 @@ class SWAttention(nn.Module):
                 varlen = attention_mask.sum(dim=1).to(torch.int32)
 
         # Call Lazy Attention Triton kernel
-        # Convert float32 parameters to model dtype for kernel computation
+        # IMPORTANT: Pass tau as float32 to ensure correct gradient accumulation in Triton
+        # Do NOT convert to q.dtype if q is bfloat16
         attn_output = lazy_attention_triton(
             q, k, v,
-            bias=self.learnable_bias_diagonals.to(q.dtype),
-            tau=self.tau.to(q.dtype),
+            bias=self.learnable_bias_diagonals,
+            tau=self.tau, # Keep as float32
             window_size=self.max_bias_length,
             varlen=varlen
         )
@@ -178,3 +179,4 @@ class SWAttention(nn.Module):
         attentions = None
 
         return o, attentions, past_key_values
+
